requirements.txt
Flask~=1.1.2
Flask-Cors~=3.0.8
marshmallow~=3.6.1
webargs~=6.1.0
Werkzeug~=1.0.1
python-dotenv